+++ changelog.md
@@ -2,6 +2,7 @@
 
 + Change example list and collection list as a function.
 + Add an iterator for grabbing examples and collections names.
++ Function name typo corrections.
 
 # Version 20.01