+++ changelog.md
@@ -3,6 +3,7 @@
 + Change example list and collection list as a function.
 + Add an iterator for grabbing examples and collections names.
 + Function name typo corrections.
++ Add "get_include" function for setup.
 
 # Version 20.01